Add recovery support for snapshot backups
Adds a SnapshotRecoveryExecutor which copies the backup label to the
destination. It has the same behaviour as other RecoveryExecutors in
that it will:

  * Mangle the config files and prepare recovery options if they are
    present in the target directory (i.e. if the config files were
    included in the snapshot).
  * Copy any WAL files into place if `--get-wal` is not used`.
  * Set the restore command if `--get-wal` is used.

It will also carry out the following checks before recovering:

  1. Verify that, For each snapshot in `backup.info`, there is a disk
     cloned from that snapshot attached to the recovery instance.
  2. Verify that each disk is mounted at the same mount point at which
     the source disk was mounted.
  3. Verify that the recovery directory is present.

Much of the logic is implemented as static methods so it can be reused
when restoring snapshot backups in Barman cloud.

Relates to #620.