[sq] Update snapshots_info support and abstract it a bit
Restructures the information in snapshots_info so that:

  1. Information which is not directly useful or necessary (i.e. which
     can't be retrieved from the GCP API at recovery time) is no longer
     included. The information recorded should be that available at
     backup time and not necessarily at recovery time, i.e. the device
     and mount info.

  2. Provider-specific information is moved into its own sub-field so
     that we can accomodate the various differences in the metadata
     when we add a second snapshot provider.

The snapshots_info is now modelled as a class which allows the
details of the attributes required by Barman (a snapshot identifier and
a device path) to be handled by a specialised SnapshotsInfo class for
each cloud provider - the calling code need only deal with accessing
`identifier` or `device` and does not need to know which
provider-specific fields these are forged from.

TODO: Tests and docstrings.