Store backup_name in backup_info at the earliest opportunity
Changes the point at which the backup_name is stored in the backup_info
to the creation of the backup_info rather than the `finally` block
in `BackupManager.backup`. Because the backup_name will not change
during the backup there is no good reason for deferring the assignment.

This fixes an issue where backup_name was not saved in the `backup.info`
file when the `--wait` option was used. This was because the backup_name
was assigned so late in the process that it relied on the final check
of the backup to save the backup info. When `--wait` is used, this
check actually happens *before* the `finally` block executes which means
the backup metadata is saved before `backup_name` is assigned. This
meant that the `--name` option was effectively ignored if the `--wait`
option was used.

Thanks to epolkerman for reporting this bug.

Closes #748.