Support SSEKMSKeyID in barman cloud uploads
Adds support for the boto3 SSEKMSKeyID parameter to barman-cloud-backup
and barman-cloud-wal-archive.

A new option `--sse-kms-key-id` is added which allows users to specify
either the key ID or the full ARN for a key stored in AWS KMS. When this
is used along with the `-e/--encryption` option, the specified key will
be used to encrypt backups and WALs when they are stored in AWS S3.

Closes #714