Disable `idle_session_timeout` for rsync-concurrent backups
Previous to this commit rsync-concurrent backups could fail
with log messages like this:

```
2023-02-20 20:18:07,846 [31148] barman.backup_executor INFO: Asking PostgreSQL server to finalize the backup.
2023-02-20 20:18:07,846 [31148] barman.backup ERROR: Backup failed issuing stop backup command (native concurrent).
DETAILS: Connection lost, reconnection not allowed
```

That issue is triggered whenever Barman copy takes more than
`idle_session_timeout` to complete because `pg_backup_start` and
`pg_backup_stop` need to be executed in the same session, and the
session is kept `idle` for as long as the copy takes.

This commit fixes that issue by disabling `idle_session_timeout`
in the PostgreSQL session where `pg_backup_start`/`pg_backup_stop`
are run.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>