Limit the rate at which new Rsync jobs are generated
This is a fix for #334 where the number of new SSH connections created
by Barman's Rsync jobs can exceed the value of MaxStartups on the target
server.

The patch works by limiting rate of *job generation* in
`copy_controller.RsyncCopyController._job_generator`. This function is
a generator function which is used by `imap_unordered` to create the
`_RsyncJob` object which is passed to the workers.

The rate is limited by tracking the job generation times within a
specifiable time window and checking if the number of jobs generated
within that window exceed a given value. Should this occur, the
generation process will wait until the oldest job start time is outside
the window before continuing.

The reason *why* this mitigates MaxStartups is that:

- The `_job_generator` generator function is lazily evaluated, so new
  jobs are not created until ready to be passed to a worker.
- `Pool.imap_unordered` uses a default chunk size of `1` which means
  jobs are passed to workers one at a time.

This means that the generation of a job in the main Barman process will
be very close in time to the execution of that job by the worker
process. So limiting the rate of job generation will also limit the rate
of job execution which will subsequently limit the rate at which new
SSH connections are made.

If the chunk size were greater than 1 or the generator function was not
lazily evaluated then this approach would not work at all (or perhaps
appear to work but only by luck).

Besides the dependence on chunk size and lazy evaluation, this approach
has a few other disadvantages:

1. It is not *directly* limiting the rate of new SSH connections. Due to
   the fact that we have no control over the OS scheduling of the worker
   processes it is possible that, even after delaying the generation of
   a new `_RsyncJob`, the actual creation of the SSH connection happens
   at the same time as the SSH connections for `_RsyncJob`s which were
   generated earlier.
2. There is no verification that existing `_RsyncJob`s have finished
   authenticating. If authentication does not complete within the time
   window used to limit the job generation rate then we will end up
   exceeding MaxStartups even though we delayed the generation of the
   next job.

Both issues above can be mitigated to some extent by increasing the
size of the window used to limit the job generation rate.

There is one significant advantage of this approach which is that it is
simple. Because rate limiting is applied in the main process we do not
have to create any new channels for communication with workers and we
do not have to worry about managing and communicating any global state
to track the number of concurrent new SSH connections. It also requires
minimal changes to the existing job controller and worker code in
Barman.

Closes #334.