Fix _copy_temporary_config_files for recovery_conf_filename
Updates `_copy_temporary_config_files` so that it copies all files
in `recovery_info["temporary_configuration_files"]` instead of just
those in `recovery_info["configuration_files"]`.

This is necessary because the recovery options file may not be an
existing configuration file (e.g. if the `--recovery-conf-filename`
option is used). In such cases it will be present in
`temporary_config_files` but not `configuration_files` and therefore
without this patch it will not be copied to the destination directory
on the recovery host.

Relates to #771.