Add support for GCP kms_key_name to barman cloud
Adds support for the google-cloud-storage kms_key_name parameter to
barman-cloud-backup and barman-cloud-wal-archive.

A new option `--kms-key-name` is added which allows users to specify
a fully qualified key name for a key stored in GCP KMS. When this is
used the specified key will be used to encrypt backups and WALs when
they are stored in GCS.

Note that if this option is used with `--snapshot-instance`, the
encryption key will only be used for the objects uploaded to GCS - that
is the `backup.info` and `backup_label` files. The encryption key cannot
be used to encrypt the snapshot. Encryption for snapshots is determined
by the encryption used on the underlying persistent disk and this is
out of Barman's control.

Closes #721