State PostgreSQL version required for --recovery-conf-filename
Updates the docs for the `--recovery-conf-filename` option so it is
clear that the option only affects recovery of backups from PostgreSQL
12 or later.

With PostgreSQL versions earlier than 12 the file for recovery
configuration options (`recovery.conf`) is used by PostgreSQL to
determine whether it is starting in recovery mode. Barman will therefore
continue to write recovery options to `recovery.conf` for PostgreSQL
versions earlier than 12 even if `--recovery-conf-filename` is
specified.

Related to #771.