[gcp-refactor] Remove knowledge of zone from barman code
Makes the snapshot zone an internal detail of the snapshot interface
so that the backup and recovery executors in barman do not need to
deal with it. This is because zone is not a meaningful parameter for
all cloud providers.

Because there are also other arguments which are specific to other cloud
providers we remove all handling of zone from barman and rely on it
being either passed in a dict of `provider_args` or being available in
the config/args. In all cases, the only code which needs to handle the
provider-specific arguments is in the factory functions which create the
interface.

This also requires moving the zone to be a property of the interface
instead of something the backup/recovery executors pass in when using
the interface. This means once the interface has been created,
everything must happen in that zone.

For backup this means the disks and instance must be in the same zone.
At recovery time the disks and instance must again be in the same
zone but this need not be the same zone that the backup instance/disks
were in - a different zone can be specified.