Rename snapshot_gcp_project option to gcp_project
Renames the config variable snapshot_gcp_project and the CLI parameter
--snapshot-gcp-project to gcp_project and --gcp-project.

This is adopting a convention which will be followed for future cloud
providers where cloud provider parameters may have a wider scope than
just snapshots. Such parameters will not be prefixed with "snapshot".

The old snapshot-prefixed options are kept for backward compatibility
but should be dropped at the next major version release of Barman.