Rename snapshot_zone option to gcp_zone
Renames the config variable snapshot_zone and the CLI parameter
--snapshot-zone-project to gcp_zone and --gcp-zone.

This is because zone is specific to gcp and the convention is that
provider-specific variables should be prefixed with the provider name.

The old snapshot-prefixed options are kept for backward compatibility
but should be dropped at the next major version release of Barman.