Rename snapshot_recovery_zone option to gcp_zone
Renames the `barman recover` and `barman-cloud-restore` argument
`--snapshot-recovery-zone` to `--gcp-zone`.

This is because zone is specific to gcp and the convention is that
provider-specific variables should be prefixed with the provider name.

This also means that the zone at recovery time overrides the zone in
the configuration and if the zone is not specified then the value from
the config is used.

To make this work we remove the `provider_args` dict and instead rely
on passing the config object or argparse namespace to the factory
functions for the snapshot interface.