Remove cloud_provider_args from test_uses_snapshot_backup_uploader
Removes the parametrized cloud_provider_args from this unit test. This
is because they are not needed for the purposes of the test. While they
would be needed in real execution, the use of the mock
get_cloud_interface factory function means that a mock is always
returned regardless of the cloud provider-specific arguments.