Delete backup_label for barman cloud snapshot backups
Deletes the backup_label from the object store when snapshot backups
are deleted. This is necessary because the existing deletion logic only
deleted the `backup.info` files and the tarballs containing the backup
data (which included the backup label). With snapshot backups, tarballs
are not used and the backup label is stored as a unique object. It must
therefore be deleted as a special case.

Closes BAR-13.