Clarify archiver and streaming_archiver defaults
Updates the default value of the `archiver` config option in the docs
with its real default and adds a note explaining:

- That it will default to `true` only if `streaming_archiver` is `off`.
- That this behaviour is deprecated and will be removed.

The docs for `streaming_archiver` are updated with a similar explanatory
note.

The streaming server configuration template is also updated so that it
instructs users to set `archiver = on` if archive_command is used in
addition to pg_receivewal.

Closes #280.