Make verification of disks optional in get_attached_volumes
Updates `CloudSnapshotInterface.get_attached_volumes` so that, when a
list of disks is supplied, the verification that those disks exist and
are attached (and the raising of a related SnapshotBackupException) can
be skipped by setting `fail_on_missing` to `False`.

This allows callers to pass in a list of disks and receive
VolumeMetadata objects only for those that are found attached to the
instance. This is useful when the calling code does additional
verification such as in the case of `find_missing_and_unmounted` disks,
where additional checks are performed to find out whether the attached
disks are mounted before the missing and unmounted disks are returned.

Previously the calling code would have to omit the disks parameter in
order to achieve this behaviour.

This commit changes no external behaviour and is preparation for adding
support for AWS snapshots, where we will need to handle the fact that
disks may be referenced either by a name or a volume ID.

Related to BAR-23.