Use region in backup_info when creating AwsCloudSnapshotInterface
Uses the region set in the `snapshots_info/provider` section of the
backup_info when creating an `AwsCloudSnapshotInterface` from a
backup_info.

This does not take precedence over the region specified in the config,
so it is possible for the user to force a different region to be used
at recovery or deletion time.