Failover handling part 1: Config
A very minimal proof of concept for cluster config.

This does not work in any scenario where WALs are not being streamed
from the same source as the backup, because Barman checks the
replication slot using the main conninfo.

We can't just use the streaming connection because that uses
streaming_conninfo to create a replication connection.