Failover handling part 2: Server implementation
Creates a WAL-specific postgres and WAL-specific streaming connection
in our Server objects. These are just references to the existing ones
if either clustering is not used or the backup host is the primary host
(since in either case we want the legacy barman behaviour). Otherwise
they are separate SQL and streaming connections to the primary.

The resulting code is far harder to comprehend than it should be so
really this commit just points to what we might want to refactor before
we do this for real.

One particularly awful bit is the merging of the replication slot status
from the WAL postgres connection into the remote status for the Server.
This is required so that Barman finds the replication slot, since it
uses the fetched status of the overall Server to do it when running
under `barman check` (when running under receive-wal it uses a direct
reference to the WAL postgres connection).