Use a Cluster as the primary abstraction for PostgreSQL servers
Adds a new class, `barman.cluster.Cluster`, which will provide an
interface through which the rest of the Barman code can access the
servers in a cluster.

Currently it just passes through to a single `barman.server.Server`
instance however there are three different ways this is references:

1. self.backup_server
2. self.wal_server
3. self.primary_server

For a non-clustered backwardly-compatible configuration, wal_server
and primary_server all point to the backup_server. Semantically this
isn't quite correct because backup_server may not be a primary_server,
however we are at early stages so we will live with this for now.

This is enough for the main bits of Barman functionality to work:

- `barman backup`
- `barman check`
- `barman receive-wal`
- `barman archive-wal`
- `barman switch-wal`
- `barman cron`

Other functions will likely be broken for now.

Also, because we allow access to `config` directly, this interface
isn't really doing its job - bits of code all over the codebase
will be accessing the config object of the backup_server directly.
Not a problem for now but when we allow backup_server to differ from
wal_server and primary_server this will be an issue.