Allow minimum chunk size to be specified
Adds a new option, `--min-chunk-size`, to `barman-cloud-backup`. If set
then this will be used when setting the minimum size of an upload chunk
in a multipart upload. The only guarantee Barman offers here is that
the upload chunk size will be greater than this value - the actual
chunk size used will be whichver is greatest of: `--min-chunk-size`,
the `MIN_CHUNK_SIZE` for the cloud provider, or the calculated chunk
size based on the `max_archive_size` and the `MAX_CHUNKS_PER_FILE`
property of the cloud interface.

This can be used to ensure uploaded chunks are not smaller than the
specified value.

Relates to BAR-113.