Do all chunk_size setting in one place
Rather than test the `--min-chunk-size` against the calculated chunk
size in one place and then test the result against the cloud interface
`MIN_CHUNK_SIZE` in another, we peform all `chunk_size` setting checks
in the `CloudUploadController` and update `CloudTarUploader` so that it
is just given the chunk size by the upload controller.

Relates to BAR-113.