Track source file for Barman configuration options
This commit introduces the class `ConfigMapping`. It extends the
functionalities of `ConfigParser` to additionally track the
configuration file which provides the effective value for each
configuration option that has been set.

That class requires that you use its `read_config` method when
attempting to parse INI files. It will take care of calling
`read`, `read_file` or `read_fp`, depending on the Python version
and on the input parameter -- if it is a file or a path. It then
exposes a `get_config_source` method to tell which configuration
file provides the effective value for a section/option.

The `get_config_source` method performs a lookup at the server
config and at the global config, because the effective server
configuration is a mix of both along with default values. If a
setting is not found in the mapping, it is assumed to be using
its default value, in which case that method returns `default`.

It is worth noting that that is how far we can go without being
more invasive because:

* As we subclass `ConfigParser`, everything else should look
  similar for Barman;
* As `ConfigParser` provides no tracking of line numbers, we
  can only track the file name with no deeper changes.

Besides that, we also add `--show-config-source` option to
`barman diagnose` command. If that flag is given by the user, then
the `config` entry of each server in the `diagnose` output is modified
so instead of simply showing the config value, it turns into a
dictionary with two keys:

* `value`: configuration value;
* `source`: source file which provides the effective configuration
  value.

References: BAR-127.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>