Add `wal_conninfo` and `wal_streaming_conninfo`
Adds the `wal_conninfo` and `wal_streaming_conninfo` configuration
values and uses them to create the `Server.wal_postgres` and
`Server.wal_streaming` PostgreSQL connections respectively.

The connection objects are created but no connection is attempted yet
because the objects are not used for anything.

The next step will be to update Barman so that `self.wal_postgres` is
used instead of `self.postgres` and `self.wal_streaming` is used instead
of `self.streaming` for activities relating to WAL streaming.