Check wal_conninfo has monitoring privileges
Adds a new field to the remote status for a PostgreSQL connection which
denotes whether a connection has monitoring privileges. This is
determined by a new `barman.postgres.PostgreSQL.has_monitoring_privileges`
function which will return if the required privileges exist for the
user *or* the user is a superuser.

These privileges allow the wal_conninfo user to check for the presence
of a replication slot and also query for replication stats.