Add --source argument to `barman replication-status`
Updates the `barman replication-status` command so that users can
specify whether replication status should be checked on the host from
which backups are sourced or the host from which WALs are sourced.

This is achieved using a new `--source` argument which can be set to
`backup-host` or `wal-host` and defaults to `backup-host`.

If backups and WALs come from the same host then the option will have
no effect and all replication information will come from the single
defined host.