wip s/snappy/cramjam/
This is a work-in-progress which attempts to replace Snappy with
Cramjam. Compression is reasonably straightforward since the
`cramjam.snappy.Compressor` really is a drop-in replacement for the
old `StreamCompressor` in python-snappy.

Decompression is a little tricky because `cramjam.snappy.Compressor`
appears to only work with complete chunks - it cannot handle being
given arbitrary chunks from a stream. We therefore copy/paste the
StreamDecompressor from python-snappy - this buffesr the incoming
data until we have a complete snappy chunk and then decompresses it.
It has been adapted so that it calls `cramjam.snappy.decompress` to
do the actual decompression.