Allow WAL streaming from a different host to the backup source
This commit makes it possible to stream WALs from a different
PostgreSQL instance than the instance used as the source for backups.

Two new configuration variables are added:

- wal_streaming_conninfo: A conninfo string which will be used by
  Barman to make streaming replication connections when receiving
  WALs.
- wal_conninfo: A conninfo string which will be used by Barman to
  manage streaming replication connections for receiving WALs.

If `wal_conninfo` is not provided then Barman will attempt to use
`wal_streaming_conninfo` for managing WAL-related replicaiton.

If `wal_streaming_conninfo` is not provided then Barman will use
the existing `streaming_conninfo` and `conninfo` strings for receiving
WALs.

The implementation works by manipulating the configuration so that
Barman connects to the WAL source when running commands which concern
WAL streaming and connects to the backup source when running all other
commands. This results in a smaller set of changes than alternative
approaches such as making the `barman.server.Server` aware of multiple
PostgreSQL instances and choosing the correct one for any given
operation.

This approach works well for the `barman backup`, `barman cron` and
`barman receive-wal` commands however both the `barman check` and
`barman replication-status` commands do need some awareness of both
connections.

For `barman replication-status`, we add a new option `--source` which
can be set to either `wal-host` or `backup-host`; this option determines
which host is queried for replication status information. The default
is `backup-host` to remain compliant with the previous behaviour.

The changes required to `barman check` (and the checks performed during
`barman backup`) are made in a subsequent commit.

Relates to BAR-120.