Improve checks when WALs and backups use different conninfos
Updates the behaviour of `barman.server.Server.check` (called during
`barman backup` and `barman check` operations) so that additional
checks are made on the connection used for streaming WALs *if* a
specific conninfo is provided for WAL streaming.

If no specific conninfo is provided for streaming WALs then the
behaviour is unchanged and all checks are carried out on the connections
specified in conninfo and streaming_conninfo.

If wal_streaming_conninfo is provided then the following checks are
carried out against the PostgreSQL host used for WAL streaming:

- "has monitoring privileges": Checks that the connection used to
  manage WAL streaming has permissions to retrieve replication slot
  information.
- "PostgreSQL streaming": Checks that the connection used for streaming
  WALs is able to connect via streaming replication.
- "wal_level": Checks that the wal_level is set to a value which allows
  physical replication on the WAL streaming host.
- "systemid coherence": Checks the system identifier of the PostgreSQL
  instance used for WAL streaming matches the system identifier known
  to Barman.
- "replication slot": Checks that the named replication slot is present
  and active on the PostgreSQL instance used for WAL streaming.

All but the last check will have already been carried out using the
conninfo and streaming_conninfo connections however it is important that
they are also performed using wal_streaming_conninfo and wal_conninfo
so that any issues with the PostgreSQL host used for WAL streaming can
be identified.

When these checks are performed for WAL streaming purposes the suffix
"(WAL streaming)" is appended to the check name.

The implementation involves extracting the logic for streaming checks
so that they can be called in isolation using remote status information
obtained from the WAL streaming connections. The following new functions
are created as a result of this:

- _check_streaming_supported
- _check_wal_level
- _check_replication_slot

In addition the existing check_identity function is updated to accept
a remote_status and suffix.

A new check, _check_has_monitoring_privileges, is also added, so that
if sufficient permissions to check the replication slot status are not
available then it will fail as a specific check rather than only as an
error case of the relpication slot check.

Relates to BAR-120.