Add support for config models
.# Description and examples

This commit adds support for config models in Barman. The main idea of this
feature is to make it easier to integrate Barman with HA solutions, e.g.
Patroni.

Besides the usual Barman server configuration, we now add support for Barman
model configuration. The model acts as a set of overrides for configuration
options for a given Barman server.

The servers configuration now have an additional configuration option:

* `cluster`: it defines the cluster name for a given Barman server. The name
  must be unique across different servers configured in Barman. If not filled,
  it defaults to the server name.

In order to say to Barman that a section of the configuration file is a model,
you need to set the configuration option `model = true` in such section. Also,
as explained for the Barman servers, you need to specify a `cluster` option in
the model. That is what ties up the Barman servers with their models.

The models currently support only overrides for these configuration options:

* `conninfo`;
* `primary_conninfo`;
* `streaming_conninfo`.

As an example, let's assume you have a PostgreSQL cluster composed of 3 nodes,
say `A` (primary), `B` (standby), and `C` (standby), and assume you are taking
backups from `A` through a Barman server named `my_cluster`. You could create
a couple config models which would override the `*conninfo` options of that
Barman server. So, in the event of a failover in the cluster, you could ask
Barman to apply a given model on top of the Barman server config, overriding
the values that would reference the failed node.

Please note that you can either have no active model, or have at most one
active model at a given point in time. If you have no active model, only the
options from the Barman server take place. If you have an active model, the
options defined in the model will override the corresponding options from the
Barman server config.

.# Technical details

Previous to this commit we used to have a `ServerConfig` class in the code.
It used to have the logic for handling the configuration of the Barman
servers. While introducing the models implementation we noticed some overlap
between how we would handle server and model configuration. With that in mind
we created a `BaseConfig` class which handles the similarities, and created a
couple subclasses from it:

* `ServerConfig`: the usual `ServerConfig` which existed before, with an
  additional `cluster` option being supported. Besides that, it now also
  tracks a few things related with the server models:
  * The hidden file `.active-model.auto.conf`, which is written under the
    Barman server directory when a model is active for that server, if any;
  * A mapping of models which are tied with this Barman server, and which
    can be activated by the user.
  * It also exposes a method `apply_model`, which will be used in a future
    commit, where we will make it possible for users to activate given
    config model for a server
  * `to_json` method has been changed, so it correctly points the source
    file for an option that is override by a model
* `ModelConfig`: a new class to handle configuration of Barman models. It
  exposes a few handy methods:
  * `get_override_options`: yields the options and values which are set by
    the model, and which override the Barman server config;
  * `to_json`: so we can put models information in the diagnose output

We had to change the logic for parsing the Barman configuration, so it is
now able to distinguish between sections for Barman servers and for Barman
models. Besides that, as we need exclusive cluster names among the Barman
servers, the function that parses the configuration also checks for that
and quits with an error if it finds conflicting names. If everything is
fine in regards to cluster names, it will append the config models for
each server near the end of parsing. At the very end it will apply any
configuration models which are active based on the hidden files that are
maintained by Barman inside each server directory.

When parsing config models, Barman will log warning messages about options
which are not supported in models. Also, it will deactivate Barman servers
if they have models which are not properly configured and that are missing
the required keys.

References: BAR-122.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>