Add documentation for HA cluster support features
This commit adds documentation for all the features that have been as part of
the work of making it easier to integrate Barman with HA clusters.

We perform the following changes to the web docs / `man`:

* Add `cluster` configuration option to `man 5`
* Add `model` configuration option to `man 5`
* Modify scope of the following configuration options in `man 5` so they
  include `model` scope:

    * `primary_conninfo`
    * `conninfo`
    * `streaming_conninfo`

* Add information about configuration models under the
  `Integration with cluster management systems` section of the web docs
* Add information about configuration models and examples under the
  `Configuration` section of the web docs
* Add documentation about the `barman config-switch` new CLI command in
  the web docs

Some changes to the documentation related with `wal_conninfo` and
`wal_streaming_conninfo` have already been applied through #868, and
are only pending to be merged.

References: BAR-136.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>