Fix servers x models relationship
Previous to this commit we had a wrong implementation of the config models: it
was considering that cluster x server was a 1-to-1 relationship, and
consequently servers x  models was a 1-to-many relationship. However, in
practice we should be able to have cluster x server as a 1-to-many relationship,
and cluster x model as a 1-to-many relationship, in such a way a model could be
used for different servers in the same cluster, resulting in a server x model
many-to-many relationship. This commit fixes the implementation in that regard.

The following changes have been performed:

* Implement `manage_model_command`, `get_model` and `get_model_list` in
  `barman.cli`: this is the equivalent feature from server handling in the CLI,
  but for models
* Move models config from `servers - SERVER -> config -> models` to `models` in
  the `barman diagnose` output: as cluster is not supposed to be a 1 to 1
  relationship with server, it makes more sense to have the models on the root
  of the output given the same model can be used by different servers
* Refactor model related functions from `ServerConfig` and `Config` classes in
  `barman.config` -- it doesn't make sense anymore to cache models in the
  `ServerConfig` object, so:
    * `ServerConfig.add_model` method has been removed
    * `ServerConfig.apply_model` method now receives a `ModelConfig` instead of
      a model name
    * `ServerConfig.to_json` has been adjusted to not expect the `models`
      attribute
    * `Config._models` will cache the models
    * `Config._populate_servers` was renamed to `_populate_servers_and_models`,
      and it is now responsible for caching the models in `Config`. Also, it
      doesn't check anymore for cluster exclusiveness among servers as that's
      not supported to be a 1-to-1 relationship

Unit tests were updated accordingly.

References: BAR-122.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>